MTP-5875: Refactor site jobs; adding cancel FN
