+++ lib/modelParser.js
@@ -146,8 +146,10 @@ export const modelParser = {
     },
     createParser(model) {
         return (data) => {
+
+            // If the response is an empty string, parse the response as an empty object.
             if(data === '') {
-                return {};
+                data = {};
             }
             const [ processedModel, processedData ] = modelParser.processModelAndData(model, data);
             const parsedObj = {};
